// app/types/auth.types.ts
export type MailType = {
  email: string;
}

export type PasswordType = {
  password: string;
}

export type LoginTypes = {
  email: string;
  password: string;
}

export interface RegisterTypes {
  firstname: string;
  lastname: string;
  dni: string;
  email: string;
  password: string;
  password2: string;
  phone: string;
}